logger: Fix crash bug when log rotate is enabled
This bug occurred in the following conditions.

* When logging to a file
* Log rotate is enabled
  * --log-rotate-threshold-size <threshold>_
  * --query-log-rotate-threshold-size <threshold>
* Process ID log output is enabled
  * --log-flags process_id

When the log is rotated, `output->file` is closed and set to NULL.
The crash occurs because of `flock(fileno(output->file), LOCK_UN);`
in that state.
This unlocking was only executed when outputting process ID to the log,
so it occurred only when outputting process ID.