ci package ubuntu: add steps for building and testing Ubuntu focal amd64's package (#1861)
GitHub: ref GH-1865

In this PR, we add steps for building and testing Ubuntu focal packages
as part of our regular development workflow on CI.
The reason for adding this is to prevent potential build failures on
Launchpad during the release process, which can result in late discovery
of issues and delays.

## Note about test using `xxHash`

Ubuntu focal provides `xxHash` version 0.7.3, which does not meet the
required version 0.8.0. We have updated the tests to skip
xxHash-dependent functionality on Ubuntu focal by adding
`@require-feature xxhash` which is the grntest's function. This change
ensures that the test about features requiring `xxHash` are executed
only when `xxHash` is available.

```
-- Checking for module 'libxxhash>=0.8.0'
--   Requested 'libxxhash >= 0.8.0' but version of xxhash is 0.7.3
You may find new versions of xxhash at https://www.xxhash.com/
-- Could NOT find GroongaxxHash (missing: GroongaxxHash_FOUND) (Required is at least version "0.8.0")
-- No xxHash
```

---------

Co-authored-by: Horimoto Yasuhiro <horimoto@clear-code.com>