packages ubuntu focal: unify dependency packages
Build (ubuntu-focal-amd64) job on CI check build error before we build package on Launchpad.

If we build packages on Launchpad, dependency packages are installed based on control.in.
On the other hand, if we build packages on CI, depndency packages are installed based on Dockerfile.

So, the dependency packages of control.in and Dockerfile msut be same.