packages ubuntu jammy: unify dependency packages (#1946)
Currently, building on Launchpad during release was a one-shot process,
and dealing with build failures is challenging.
To mitigate this, we're setting up a CI environment that mimics the
Launchpad environment for testing.

However, the dependency packages listed in `control.in` and `Dockerfile`
are different, making the test ineffective. This change unifies the
dependency packages in `Dockerfile` to match those in `control.in` to
ensure consistency between the test and actual build environments.