request_cancel: Fix a bug that canceled request may return an invalid response (#2013)
Sometimes the following response is received when canceling.

Example of command version 3 (Incomplete JSON) (`...` is the omission of
an element):

```
{
  "body": {
  "columns":[...],
  "records":[
    [...],
    "header":{
      "return_code":-77,
      ...
    }
  }
```

If `ctx->rc` is `GRN_CANCEL`, we force using an empty result. Because
creating result may be incomplete.
We can do it because the cancelled response must not be used.