grn_expr_executor_add_replace_value: add
In general, expr_executor uses values that are stored in column. We
can use the specified value instead of stored value.

It's useful for generated column. We'll use existing hook mechanism
for generated column. It's called BEFORE we set a column value. So
stored value isn't what we want to use.

We can use the specified value instead of stored value for building
generated column by hook with this.