request_cancel: Fix a bug that canceled request may return an invalid response
Sometimes the following response is received when canceling.
Example of command version 3 (Incomplete JSON) (`...` is the omission of an element):

```
{
  "body": {
  "columns":[...],
  "records":[
    [...],
    "header":{
      "return_code":-77,
      ...
    }
  }
```

I found that the e5400b36db0820c8baf37bacafa7d70b67129bb9 approach did not fix it.
Bugs with invalid JSON occur during the streaming process.

In particular, `logical_range_filter` may raise an exception and not output all if `request_cancel`.
Fix `logical_range_filter` to output closing brackets in case of exception due to cancellation, so that it becomes normal JSON.