output: Add grn_output_result_set_content()
GitHub: Related: GH-2031

Prepare to fix the output of `records`.

Add grn_output_result_set_content() to output `[...]` of `records`.

Added `grn_output_table_records_close_v*` but removed it
because it was already abstracted and unnecessary.

`grn_output_table_records_content_v*` was removed since it is common in v1 and v3.