doc conf: allow include directives to contain Groonga command as a content
This change resolves warnings during documentation generation related to
the `include` directive containing content. By default, the include directive
does not permit any content, which resulted in warnings like:

```console
$ cmake \
    -S .
    -B ../groonga.doc
    --preset=doc && \
  cmake
    --build ../groonga.doc
...
/home/otegami/work/c/groonga/doc/source/reference/functions/language_model_vectorize.md.rst:61: WARNING: 'include': Has content, but none permitted [myst.directive_parse]
...
```

In the Groonga project, we need to embed the results of executing
Groonga commands within include directives. To achieve this, we
modify the default behavior of the `include` directive to allow it
to contain content. This prevents the warnings and facilitates the
integration of Groonga command outputs into the documentation.