doc conf: suppress `'include': Has content` warnings in .md (#2054)
We want to embed the results of executing Groonga commands within
documents for easy to understand. We preprocess our document sources by
`doc/update-execution-example.rb` and generate execution results.

We're using the following syntax for the preprocess in `.md`:

````md
<!-- groonga-command -->

```{include} output-path.md
table_create ...
select ...
```
````

But this causes the following warnings:

```console
$ cmake \
    -S .
    -B ../groonga.doc
    --preset=doc && \
  cmake
    --build ../groonga.doc
...
/home/otegami/work/c/groonga/doc/source/reference/functions/language_model_vectorize.md.rst:61: WARNING: 'include': Has content, but none permitted [myst.directive_parse]
...
```

We can use other syntax for `.md` because
`doc/update-execution-example.rb` is maintained by us. We can just
implement a logic for other syntax in `doc/update-execution-example.rb`.

FYI: We're using the following syntax in `.rst`:

```rst
.. groonga-command
.. include:: output-path.log
.. table_create ...
.. select ...
```

For example, `.rst` uses comment (`.. ` is a comment syntax in `.rst`)
for Groonga commands. We may be able to use `% table_create ...` in
`.md` but we don't have a good idea for new syntax for now.

This change suppresses the warning by changing the `include` directive's
property. MyST-Parser generates the warning for no content directives
such as the `include` directive. This changes the `include` directive to
a have content directive.