NormalizerNFKC: Add support for `unify_middle_dot` and `remove_symbol` combination (#2080)
GitHub: fixes GH-1538

When `unify_middle_dot` is enabled, the middle dot becomes a symbol and
should be removed with `remove_symbol`.

This should be implemented as normalization -> remove symbol order but
the current implementation is difficult to use the approach. So we check
middle dot unification in remove symbol phase for now.