pat: Fix the calculation for the target segment
When the next length to be added equals `GRN_PAT_SEGMENT_SIZE`, move on to the next segment.
But, if the size is the equal, it can be written to the current segment, so it was fixed so that it can be calculated correctly.

The io layer allocates an space of `GRN_PAT_SEGMENT_SIZE`.