delete: add support for referenced weight vector
If a referenced record is deleted, values that refer the record are
also removed when there are indexes for them.

It worked for non-weight vector column but didn't work for weight
vector column. If the feature was used for weight vector column, the
value of the weight vector column will be broken.

This fixes the problem.