doc: Prepare to add NormalizerNFKC reference
Since the content is almost the same, `NormalizerNFKC` documentation will be based on `NormalizerNFKC150`.
To make it easier to see what NormalizerNFKC-specific changes were made to the NormalizerNFKC150 document,
we will first commit the NormalizerNFKC150 document as is.

This commit copied `normalizer_nfkc150.rst` to `normalizer_nfkc.rst`.
(However, duplicate anchors will cause an error, so we have replaced
some of them, such as `.. _normalizer-nfkc150` => `.. _normalizer-nfkc`.)