ii parallel-build: don't reuse BlockBuilder
If we reuse BlockBuilders, it's difficult to manage child grn_ctxs.

The current implementation may be broken when the target lexicon uses
normalizer/tokenizer/token-filter options. Because options for
temporary tables keep in grn_ctx. If we can't align life-time of a
BlockBuilder (that has a temporary lexicon) and a child grn_ctx,
we may not be able to get correct options.

We can fix the problem by avoiding re-using BlockBuilder.