test: command_line: Fix encoding conversion of logs (#2267)
On Windows, `log_path.read.encoding` is UTF-8, but the characters may be
broken.
In such a case, conversion with `encode("UTF-8", "Windows-31J")` will
result in the expected encoding.

Even if the encoding of the log is UTF-8, the result of `encode("UTF-8",
"Windows-31J")` will not be broken.