Fix Apache Arrow output for slice and drilldown
When running slice or drilldown with
output_type=apache-arrow and command_version=3,
Groonga would crash because it attempted to set a
label in the Arrow output, which isn’t supported in
that context.

To prevent this, we added `GROONGA:data_type` and
`GROONGA:label` metadata to the Apache Arrow output
for slice and drilldown results when using
`command_version=3`.