tokenize: add new `--output_format` option
Currently, the output of tokenize command show
tokenized values includes additional information.
This output is useful for debugging. However, this
output is hard to read when users want to know the
tokenized token values.

So, we add a new option to output only tokenized token
values. It will get easier for users to know only
tokenized token values.