pat: expand max of the number of pat segments when we specify KEY_LARGE flag
This commit is part of the work to implement #2349.

We increase the number of pat segments when we use `KEY_LARGE` flag by this commit.
This is prepare for We can expand max total key size to 4TiB from 4GiB when we use `KEY_LARGE` flag.

Supports of Read from/Write to Expanded segments implement in theother commits.