Add a new pat node struct for using KEY_LARGE
This commit is part of the work to implement #2349.

We add a new struct for pat node has large key.
Because we don't allocate a needless memory space when we don't use `KEY_LARGE` flag.

We implement only definition of a new struct and a new union in this commit.
We will implemnt codes use this struct and union by the other commit.