pat: expand max number of pat segments when we specify KEY_LARGE flag (#2373)
This commit is part of the work to implement
https://github.com/groonga/groonga/issues/2349.

We increase the number of pat segments when we use `KEY_LARGE` flag by
this commit.
This is prepare for We can expand max total key size to 1TiB from 4GiB
when we use `KEY_LARGE` flag.

Supports of Read from/Write to Expanded segments implement in the other
commits.

---------

Co-authored-by: Sutou Kouhei <kou@cozmixng.org>