ii near-phrase/near-phrase-product: fix interval caclulation (#2411)
This problem is similar problem to GH-2407.

Currently, we can not calculate correctly interval between phrase in
`*NP` and `*NPP` operator too.
`*N` operator also include this problem. However, `*N` operator doesn't
occur this problem basiaully.
Because `*N` doesn't process phrases basically.


We should use the number of tokens in phrase of previous phrase in `*NP`
and `*NPP` operator too.
However, `*NP` and `*NPP` operator use the number of tokens in phrase of
current phrase currently in caculating interval between phrases.

So, we modify that use it of previous phrase in this commit.