pat: use pat_node_common instead of pat_node in grn_pat_del_data structure  (#2424)
This commit is part of the work to implement GH-2349.

I added the following function temporary.

- _grn_pat_wal_add_entry_data_set_record_direction()
- _pat_node_get_key()
- _grn_pat_next_location()

Because if we modify the above functions, we must modify many place in
lib/pat.c to prevent build error.
So, I added the above functions temporary to reduce affect.

The above functions will change to functions that have no "_" prefix at
the completion of GH-2349.