pat: use pat_node_common instead of pat_node in defrag (#2425)
This commit is part of the work to implement GH-2349.

I added "_pat_get()" function temporary.
Because if we modify the above functions, we must modify many place in
lib/pat.c to prevent build error.
So, I added the above functions temporary to reduce affect.

The above functions will change to functions that have no "_" prefix at
the completion of GH-2349.