ii ordered-near-phrase-product: fix interval caclulation (#2442)
This problem is similar problem to GH-2407.

Currently, we can not calculate correctly interval between phrase in
`*ONPP` operator too.

We should use the number of tokens in phrase of previous phrase in
`*ONPP` operator too.
However, `*ONPP` operator use the number of tokens in phrase of current
phrase in caculating interval between phrases.

So, we modify that use it of previous phrase in this commit.