cmake: Enable C4244 when GRN_WARN_CONVERSION=ON
C4244 is also a "possible loss of data" warning as well as C4267, so add it.

C4244: 'argument' : conversion from 'type1' to 'type2', possible loss of data
C4267: 'var' : conversion from 'size_t' to 'type', possible loss of data

Add `/W3` because the default is `/W1`, so level2 and level3 are not warnings.

https://learn.microsoft.com/en-us/cpp/build/reference/compiler-option-warning-level
> /W1 displays level 1 (severe) warnings. /W1 is the default setting in the command-line compiler.