cmake: suppress declaration-after-statement warning for llama.cpp (#2461)
## Issue

We got the following error when we built Groonga with MariaDB.

```
/home/buildbot/_deps/llama_cpp-src/ggml/src/ggml.c: In function 'ggml_quantize_chunk':
/home/buildbot/_deps/llama_cpp-src/ggml/src/ggml.c:6405:5: error: ISO C90 forbids mixed declarations and code [-Werror=declaration-after-statement]
 6405 |     const size_t start_row = start / n_per_row;
      |     ^~~~~
cc1: all warnings being treated as errors
```

## Cause

When building Groonga with MariaDB, the bundled
llama.cpp triggers `-Werror`.

## Solution

Upstream llama.cpp doesn't use C99 standardeven since this project is
new project. It is unlikely to
change this style, so we suppress the warning by
explicitly passing -Wno-declaration-after-statement in Debug and
RelWithDebInfo builds.

---------

Co-authored-by: Abe Tomoaki <abe@enzou.tokyo>