table: use the header of table object instead of the header of grn_obj (#2486)
Fix: GH-2484

Groonga refers to the header of grn_obj when executing a truncate
command.
However, the header of `grn_obj` does not include the flag information
of individual tables.

Currently, Groonga cannot set the correct flags when executing a
truncate command on a table.
As a result, `TABLE_PAT_KEY` and `TABLE_HASH_KEY` lose the `KEY_LARGE`
flag after the truncate command is executed.

With this modification, Groonga now refers to the header of each
individual table.