cmake: suppress declaration-after-statement warning for BASE64
We got the following error when we built Groonga with MariaDB.

```
/home/runner/work/mroonga/mroonga/mariadb.build/_deps/base64-src/lib/arch/generic/64/enc_loop.c: In function ‘enc_loop_generic_64_inner’:
/home/runner/work/mroonga/mroonga/mariadb.build/_deps/base64-src/lib/arch/generic/64/enc_loop.c:15:9: error: ISO C90 forbids mixed declarations and code [-Werror=declaration-after-statement]
   15 |         const size_t index0 = (src >> 52) & 0xFFFU;
      |         ^~~~~
```

When building Groonga with MariaDB, the bundled
base64 triggers `-Werror`.

Upstream base64 doesn't use C99 or earlier standardeven. It is unlikely to change this style, so we suppress the warning by
explicitly passing -Wno-declaration-after-statement in Debug and RelWithDebInfo builds.