compressor: add (#2559)
It abstracts compression routines for zlib, LZ4, Zstandard and Blosc2.

grn_ja uses this instead of raw zlib, LZ4, Zstandard and Blosc2 API.

grn_ja_reader is also changed but it may not work. We will revisit this
when we use grn_ja_reader. There is no grn_ja_reader user for now. We
need streaming API and cache API when we revisit this.

This is for adding new compression routine such as OpenZL and adding
support for table key compression.