cmake: enable -Wtype-limits warning and fix comparison
## Issue

We got the following error when we built Groonga with MariaDB.

```
/home/buildbot/extra/groonga/lib/raw_string.c:132:24: error: comparison of unsigned expression in '< 0' is always false [-Werror=type-limits]
```

### Hot to reproduce

We can reproduce it by specifying `-Werror=type-limits`.
```console
cmake \
  -S . \
  -B ../groonga.build \
  --preset=debug-default \
  -DCMAKE_INSTALL_PREFIX=/tmp/local \
  -DGRN_WITH_APACHE_ARROW=ON \
  -DCMAKE_C_FLAGS="-Werror=type-limits" && \
cmake --build ../groonga.build/
```

## Causes

When building Groonga with MariaDB, the bundled llama.cpp triggers `-Werror`.

## Solution

This PR fixes the existing violation in
`grn_raw_string_substring`() where `size_t` was
compared with < 0, which is always false since
`size_t` is unsigned.

And also, it enables the `-Wtype-limits` compiler
warning to catch comparisons that are always true or
false due to the type's range.